\name{centerDat}
\docType{data}
\alias{centerDat}
\title{Center Data set}
\description{
centerDat is data frame contains following information
Center       as int  1 2 3 4 5 9
Location     as Factor like "Amsterdam","Berlin"
NumOfSubjects as int  4 2 1 1 3 2
Latitude     as num  like 51.5 48.8 52.4 52.5 41.6 ...
Longtitude   as num similar to the Latitude  
}
\usage{
data("dmRToolsData")
centerDat}
\format{data frame}
\author{ZekaiOtles}
\keyword{datasets}