\name{crfsDlyDat}
\docType{data}
\alias{crfsDlyDat}
\title{incoming CRFs Data set for daily}
\description{
incoming crf information comes in a matrix format with following column headers
month,day,year, weeks,incoming.CRF.pages
month as int 1..12
day as int 1..31
weeks as int 1..54
incoming.CRF.pages as int
}
\usage{
data("dmRToolsData")
crfsDlyDat}
\format{data matrix}
\author{ZekaiOtles}
\keyword{datasets}
