\name{crfsMonDat}
\docType{data}
\alias{crfsMonDat}
\title{incoming CRFs Data set for Monthly}
\description{
incoming crf information comes in a matrix format with following column headers
month,day,year, weeks,incoming.CRF.pages
month as int 1..12
day as int 1..31
weeks as int 1..54
incoming.CRF.pages as int
}
\usage{
data("dmRToolsData")
crfsMonDat}
\format{data matrix}
\author{ZekaiOtles}
\keyword{datasets}
