\name{ctQcsDat}
\docType{data}
\alias{ctQcsDat}
\title{center Qc Data set}
\description{
Center Qcs data has following columns; 
Site_Number,Num_of_RECORDS,Num_of_QCS,Num_Resolved_QCs,
QCage1Wk,QCage2Wks,QCage1Month,QCage2Months,QCageGT2Months 
center Qcs data comes in a matrix format with following column headers
all the columns data comes as int

Also QCage column headers could be modified to reflect the actual data
}
\usage{
data("dmRToolsData")
ctQcsDat}
\format{data matrix}
\author{ZekaiOtles}
\keyword{datasets}