\name{dmRlegScales}
\alias{dmRlegScales}
\title{Scales Used in Legends Function}  
\description{
  Intent to use for scaling parameters passing to the figures
}
\usage{
dmRlegScales(par)
}
\arguments{
\item{par}{par is scaling parameter}
}
\seealso{
 \code{\link{dmRlegend}}
}
\examples{
library("dmRTools")
data("dmRToolsData")
scale<-dmRlegScales(par) }

\author{Zekai Otles}
