% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DMRtools.R
\docType{package}
\name{DMRtools-package}
\alias{DMRtools}
\alias{DMRtools-package}
\title{Helper functions for differential methylation analysis for 450K array}
\description{
Helper functions for differential methylation analysis for 450K array
}

