% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg.DMR.plot.R
\name{gg.DMR.plot}
\alias{gg.DMR.plot}
\title{Plot methylation pattern in genomic region}
\usage{
gg.DMR.plot(region = NULL, gene = NULL, f = fDat, ...)
}
\arguments{
\item{region}{A character value formatted e.g. "chr5:122433740-122435550"}

\item{gene}{A gene symbol to plot}

\item{f}{A data.frame of CpG site annotations for the 450K annotation, which is subsetted for the \code{gene.plot} function}

\item{...}{Arguments to pass}
}
\value{
ggplot object, same as output of \code{gene.plot}
}
\description{
Plot methylation pattern in genomic region
}
\seealso{
\code{\link{gene.plot}}
}

