% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limma.tt.R
\name{limma.tt}
\alias{limma.tt}
\alias{limma.tt.paired}
\title{Generate limma top-table for differential methylation}
\usage{
limma.tt(mval.mat, betas.mat, pheno, var.interest, covars = NULL, fDat)

limma.tt.paired(mval.mat, betas.mat, pheno, subj.fld, var.interest,
  covars.diff = NULL, covars.fixed = NULL, fDat)
}
\arguments{
\item{mval.mat}{Matrix of M-values with CpG sites in rows and samples in columns}

\item{betas.mat}{Matrix of \eqn{\beta}-values corresponding to \code{mval.mat}}

\item{pheno}{Data frame of phenotypes describing samples in same order as columns of \code{mval.mat} and \code{betas.mat}}

\item{var.interest}{Character value for one column name in \code{pheno}. For \code{limma.tt}, this is simply the variable in the linear model on which limma will conduct inference. For \code{limma.tt.paired}, this variable must be a two-level factor for which each subject has two samples in the array, one for each factor level.}

\item{covars}{Character vector of covariates to include in the linear model. These should all be names of columns in \code{pheno}.}

\item{fDat}{Data.frame of CpG site annotations. Must contain columns named "GeneSymbol" and "cpgSite".}

\item{subj.fld}{Character value for the ID of subjects in the experiment. Must correspond to a column name of \code{pheno}. For \code{limma.tt.paired}, each different subject must have exactly two samples in the array, one for each level of \code{var.interest}.}

\item{covars.diff}{Character vector of covariates which vary according to the levels of \code{var.interest} in the paired analysis. The differences of these covariates will be added to the linear model. Must correspond to columns of \code{pheno}.}

\item{covars.fixed}{Character vector of covariates which do not vary according to the levels of \code{var.interest}. These will be added as is to the linear model in the paired analysis. Must correspond to columns of \code{pheno}.}
}
\value{
A data.frame giving the standard output from a limma analysis for the M-values, as well as a column denoting effect sizes for corresponding \eqn{\beta}-values, called \code{betaFC}, and a column of corresponding genes for the CpG sites.
}
\description{
Creates a custom output for testing differential methylation of M-values
according to a variable within a linear model for each CpG site in a 450K array.
}
\details{
\code{limma.tt()} Runs limma on M-values based on a linear model including a variable of interest (for inference) and covariates.
\code{limma.tt.paired()} Is a custom function which takes in an array of M-values of paired samples. In this analysis, the variable of interest is a two-level factor and differences in M-values are calculated per subject according to this factor. Covariates are included in the linear model as either differences according to the variable of interest or fixed covariates. The inference is then done on the intercept in the linear model to test for differences according to the variable of interest per subject, adjusting for appropriate covariates.
}
\examples{

}

