\name{qcsPerPlateDat}
\docType{data}
\alias{qcsPerPlateDat}
\title{Qc per CRF or plate Data set}
\description{
Qcs per plate or CRFs dataset which has following columns;
CRFNAME,FIELDNAME,DESCRIPTION,MISS,RANG,       
INCONS,ILLEGIB,FAXN,OTHR
all the columns except description data comes as int
}
\usage{
data("dmRToolsData")
qcsPerPlateDat}
\format{data matrix}
\author{ZekaiOtles}
\keyword{datasets}