% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset.dmrcate.R
\name{subset.dmrcate}
\alias{subset.dmrcate}
\title{subset DMRcate output}
\usage{
\method{subset}{dmrcate}(res.obj, rmBlankAssoc = T, min.no.probes = 5)
}
\arguments{
\item{res.obj}{A \code{dmrcate.output} object, generated by \code{DMRcate::dmrcate()}}

\item{rmBlankAssoc}{Logical. If \code{TRUE}, removes regions that have no known genes in the region}

\item{min.no.probes}{Integer specifying the minimum number of CpG sites per DMR to keep in the results object (must be at least 1)}
}
\description{
filter differentially methylated regions for gene annotation and minimum number of probes
}

