% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize.dmrcate.R
\name{summarize.dmrcate}
\alias{summarize.dmrcate}
\title{Summarize differential methylation analysis}
\usage{
summarize.dmrcate(res.obj)
}
\arguments{
\item{res.obj}{\code{dmrcate.output} object output from \code{\link{DMRcate::dmrcate}}.}
}
\value{
named vector of length 3.
"min3":  number of DMRs with a minimum of 3 CpG sites
"min3_beta02": number of DMRs with at least 3 CpG sites and a max effect size of \eqn{|\beta| > 0.2}
"min3_anno": number of DMRs with at least 3 CpG sites and at least one gene annotated to the region.

 @export
}
\description{
Create a brief numerical based on the output of a \code{DMRcate} analysis of differentially methylated regions (DMRs)
}

