% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize.tt.cutoffs.R
\name{summarize.tt.cutoffs}
\alias{summarize.tt.cutoffs}
\title{Summarize top-table for various cutoffs}
\usage{
summarize.tt.cutoffs(tt, pval.thresh, beta.thresh = c(0.05, 0.1, 0.2, 0.3,
  0.4, 0.5))
}
\arguments{
\item{tt}{A top-table output from \code{limma}}

\item{pval.thresh}{Numerical vector of p-values in (0,1)}

\item{beta.thresh}{Numerical vector of contrast (effect) sizes}
}
\value{
Matrix of integers for each contrast and p-value pair, how many features have a lower p-value and contrast of greater magnitude
}
\description{
Get counts from a limma analysis of how many features pass statistical significance and effect-size relevance
}

