\name{visScheduleDat}
\docType{data}
\alias{visScheduleDat}
\title{Scheduled Study Data }
\description{
This is actual study data with all the information containg the scheduled visit for the subjects
it contains following variables
CRF# as int
visitNo    as  int  
SubjectID  as int  
SubjectInit  as characters
VisitDate as date
}
\usage{
data("dmRToolsData")
visScheduleDat}
\format{data frame}
\author{ZekaiOtles}
\keyword{datasets}