% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volcano_plot.DMRcate.R
\name{volcano_plot.DMRcate}
\alias{volcano_plot.DMRcate}
\title{Plot for a DMRcate output}
\usage{
volcano_plot.DMRcate(res.obj, pval.fld = c("mean", "min"), pthresh = 0.01,
  contrast.thresh = 0.25, min.p = 1e-100, xlabel = "max beta difference",
  ylabel = "-log10 p-value", title = "", label.thresh = 0.001,
  colors = c(neut = "darkgrey", low = "red", high = "blue"))
}
\arguments{
\item{res.obj}{A \code{dmrcate.output} object}

\item{pval.fld}{Which p-value from DMRcate analysis to plot, either the mean p-value in the DMR (\code{'mean'}) or the minimum p-value (\code{'min'}).}

\item{pthresh}{Numeric value signifying the maximum p-value at which DMRs should be colored}

\item{contrast.thresh}{Numeric value signifying the threshold for maximum absolute effect size at which DMRs should be colored.}

\item{min.p}{Numeric minimum value for plotted p-values. Gives an upper limit to the y-axis}

\item{xlabel, ylabel, title}{Text labels for axis and main titles in plot}

\item{label.thresh}{Value for maximum p-value threshold with which to label points with gene names}

\item{colors}{for neutral, hypomethylated, and hypermethylated regions, respectively}
}
\value{
A \code{ggplot} object.
}
\description{
Volcano-style plot of effect-size vs. significance with size of DMR represented
}
\details{
Each DMR in the input object will be colored according to whether they are considered hypermethylated in the reference group, hypomethylated, or neither. This will be determined by user-specified p-value and effect-size thresholds
}

